requirements.txt
Flask==2.2.5
Werkzeug==2.2.3
Flask-Caching==2.1.0
Flask_Dance==7.0.0
Flask-Mail==0.9.1
Flask-Redis==0.4.0
PyYAML==6.0
python-dateutil==2.8.2
requests==2.31.0
cryptography==41.0.3
packaging==21.3
sqlalchemy==1.4.32
SQLAlchemy-Utils==0.38.2
Flask-SQLAlchemy==2.5.1
Flask-Migrate==3.1.0
Flask-Alembic==2.0.1
alembic==1.7.7
Flask-Script==2.0.6
tzlocal==4.1
PyMySQL==1.0.2
python-swiftclient==3.13.1
python-keystoneclient==4.4.0
pycryptodomex==3.14.1
hvac==0.11.2
MarkupSafe==2.1.1
boto3==1.21.27
python-ldap==3.4.0
randomcolor
sshpubkeys
